requirements.txt
pillow==9.2.0
requests==2.28.1
geopy==2.2.0

# Projects required for Django:
django==4.0.6
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
django-filter==22.1
django-ipware==4.0.2
mysqlclient==2.1.1
PyJWT==2.4.0

# AWS EC2 imports:
wheel
gunicorn==20.0.2
